# fct-frontend/api-spec/openapi.yaml
openapi: 3.1.1
info:
    title: REST API for the website of "Academia Fabiana Perez"
    description: REST API for the website of "Academia Fabiana Perez"
    version: 1.0.0
servers:
    - url: http://localhost:8080/api/v1
paths:
...
    /courses/img/{id}:
        put:
            summary: Update course image
            operationId: updateCourseImage
            tags:
                - Courses
            parameters:
                - name: id
                  in: path
                  description: Course id
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                description: Course image to update
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                img:
                                    type: string
                                    format: binary
                            required: [img]
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Course"
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ValidationErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Course not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        delete:
            summary: Delete course image
            operationId: removeCourseImage
            tags:
                - Courses
            parameters:
                - name: id
                  in: path
                  description: Course id
                  required: true
                  schema:
                      type: integer
                      format: int64
                - name: imgName
                  in: query
                  description: Image name
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    description: Successful operation
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ValidationErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Course not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
...
    /products:
        get:
            summary: Get all products
            operationId: getAllProducts
            tags:
                - Products
            parameters:
                - name: productCategoryId
                  in: query
                  description: Filter by product category id
                  required: false
                  schema:
                      type: integer
                      format: int64
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Product"
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        post:
            summary: Create a new product
            operationId: createProduct
            tags:
                - Products
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/AddProductRequest"
            responses:
                201:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Product"
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                409:
                    description: Product already exists
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
    /products/{id}:
        get:
            summary: Get a product by id
            operationId: getProductById
            tags:
                - Products
            parameters:
                - name: id
                  in: path
                  description: Product id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Product"
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Product not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        put:
            summary: Update a product
            operationId: updateProduct
            tags:
                - Products
            parameters:
                - name: id
                  in: path
                  description: Product id
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                description: Product to update
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateProductRequest"
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Product"
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Product not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        delete:
            summary: Delete a product
            operationId: deleteProduct
            tags:
                - Products
            parameters:
                - name: id
                  in: path
                  description: Product id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                204:
                    description: Successful operation
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Product not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
...
    /products/img/{id}:
        put:
            summary: Update product image
            operationId: updateProductImage
            tags:
                - Products
            parameters:
                - name: id
                  in: path
                  description: Product id
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                description: Product image to update
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                img:
                                    type: string
                                    format: binary
                            required: [img]
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Product"
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Product not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        delete:
            summary: Delete product image
            operationId: removeProductImage
            tags:
                - Products
            parameters:
                - name: id
                  in: path
                  description: Product id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                200:
                    description: Successful operation
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Product not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
...
components:
    schemas:
...
        Course:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                startDate:
                    type: string
                    format: date
                endDate:
                    type: string
                    format: date
                enrollmentPrice:
                    type: number
                    format: double
                categoryId:
                    type: integer
                    format: int64
                description:
                    type: string
                imgNames:
                    type: array
                    items:
                        type: string
            required:
                [
                    id,
                    startDate,
                    endDate,
                    enrollmentPrice,
                    categoryId,
                    description,
                    imgNames,
                ]
...
        Product:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
                description:
                    type: string
                price:
                    type: number
                    format: double
                stock:
                    type: integer
                    format: int32
                imageName:
                    type: string
                productCategory:
                    $ref: "#/components/schemas/ProductCategory"
            required:
                [
                    id,
                    name,
                    description,
                    price,
                    stock,
                    imageName,
                    productCategory,
                ]
        ProductCategory:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
                vatPercentage:
                    type: number
                    format: double
            required: [id, name, vatPercentage]
...
        AddProductRequest:
            type: object
            properties:
                name:
                    type: string
                    x-field-extra-annotation: '@NotBlank(message = "Name field cannot be blank")'
                description:
                    type: string
                    x-field-extra-annotation: '@NotBlank(message = "Description field cannot be blank")'
                price:
                    type: number
                    format: double
                    minimum: 1
                stock:
                    type: integer
                    format: int32
                    minimum: 1
                productCategoryId:
                    type: integer
                    format: int64
            required:
                [name, description, price, stock, imageName, productCategoryId]
        UpdateProductRequest:
            type: object
            properties:
                name:
                    type: string
                    x-field-extra-annotation: '@NotEmpty(message = "Name field cannot be empty")'
                description:
                    type: string
                    x-field-extra-annotation: '@NotEmpty(message = "Description field cannot be empty")'
                price:
                    type: number
                    format: double
                    minimum: 1
                stock:
                    type: integer
                    format: int32
                    minimum: 1
                productCategoryId:
                    type: integer
                    format: int64
...
        ErrorMessage:
            type: object
            properties:
                timestamp:
                    type: string
                    format: date-time
                    example: "1970-01-01T00:00+18:00"
                status:
                    type: integer
                    format: int32
                    example: 404
                error:
                    type: string
                    example: "Not Found"
                message:
                    type: string
                    example: "Resource not found"
                path:
                    type: string
                    example: "/api/v1/resource/1"
            required: [timestamp, status, error, message, path]
        ValidationErrorMessage:
            type: object
            properties:
                timestamp:
                    type: string
                    format: date-time
                    example: "1970-01-01T00:00+18:00"
                status:
                    type: integer
                    format: int32
                    example: 404
                error:
                    type: string
                    example: "Bad Request"
                messages:
                    type: object
                    additionalProperties:
                        type: string
                    example:
                        "name": "Name field cannot be blank"
                path:
                    type: string
                    example: "/api/v1/resource/1"
            required: [timestamp, status, error, messages, path]
    securitySchemes:
        BearerAuth:
            description: Bearer token using a JWT
            type: http
            scheme: bearer
            bearerFormat: JWT